chore(tests): move check_metrics next to CountdownPlugin
