+++ falco_plugin_tests/tests/source_single.rs
@@ -2,26 +2,11 @@
 mod tests {
     use falco_plugin::base::Plugin;
     use falco_plugin_tests::plugin_collection::source::countdown::{
-        CountdownPlugin, COUNTDOWN_PLUGIN_API,
+        check_metrics, CountdownPlugin, COUNTDOWN_PLUGIN_API,
     };
     use falco_plugin_tests::{
         init_plugin, instantiate_tests, CapturingTestDriver, PlatformData, ScapStatus, TestDriver,
     };
-
-    fn check_metrics<C: CapturingTestDriver>(driver: &mut C, batches: usize, events: usize) {
-        let metrics = driver.get_metrics().unwrap();
-        let mut metrics = metrics.iter();
-
-        let m = metrics.next().unwrap();
-        assert_eq!(m.name, "countdown.next_batch_call_count");
-        assert_eq!(m.value, batches as u64);
-
-        let m = metrics.next().unwrap();
-        assert_eq!(m.name, "countdown.events_produced");
-        assert_eq!(m.value, events as u64);
-
-        assert!(metrics.next().is_none());
-    }
 
     fn test_dummy_next<D: TestDriver>() {
         let (driver, _plugin) = init_plugin::<D>(
